--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -14,7 +14,8 @@ function N3Store(triples, options) {
   // `_entities` maps entities such as `http://xmlns.com/foaf/0.1/name` to numbers.
   // This saves memory, since only the numbers have to be stored in `_graphs`.
   this._entities = Object.create(null);
-  this._entities['><'] = 0; // Dummy entry, so the first actual key is non-zero
+  // `_entitiesInverse` is the inverse mapping of `_entities`, this is to avoid expensive inverse lookups.
+  this._entitiesInverse = Object.create(null);
   this._entityCount = 0;
   // `_blankNodeIndex` is the index of the last created blank node that was automatically named
   this._blankNodeIndex = 0;
@@ -88,7 +89,7 @@ N3Store.prototype = {
   // (for instance: _subject_, _predicate_, and _object_).
   // Finally, `graph` will be the graph of the created triples.
   _findInIndex: function (index0, key0, key1, key2, name0, name1, name2, graph) {
-    var results = [], entityKeys = Object.keys(this._entities), tmp, index1, index2;
+    var results = [], entityKeys = this._entitiesInverse, tmp, index1, index2;
 
     // If a key is specified, use only that part of index 0.
     if (key0) (tmp = index0, index0 = {})[key0] = tmp[key0];
@@ -169,9 +170,10 @@ N3Store.prototype = {
     // Instead, we have a separate index that maps entities to numbers,
     // which are then used as keys in the other indexes.
     var entities = this._entities;
-    subject   = entities[subject]   || (entities[subject]   = ++this._entityCount);
-    predicate = entities[predicate] || (entities[predicate] = ++this._entityCount);
-    object    = entities[object]    || (entities[object]    = ++this._entityCount);
+    var entitiesInverse = this._entitiesInverse;
+    subject   = entities[subject]   || (entities[entitiesInverse[++this._entityCount] = subject]   = this._entityCount);
+    predicate = entities[predicate] || (entities[entitiesInverse[++this._entityCount] = predicate] = this._entityCount);
+    object    = entities[object]    || (entities[entitiesInverse[++this._entityCount] = object]    = this._entityCount);
 
     var changed = this._addToIndex(graphItem.subjects,   subject,   predicate, object);
     this._addToIndex(graphItem.predicates, predicate, object,    subject);