--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -14,6 +14,7 @@ function N3Store(triples, options) {
   // `_entities` maps entities such as `http://xmlns.com/foaf/0.1/name` to numbers.
   // This saves memory, since only the numbers have to be stored in `_graphs`.
   this._entities = Object.create(null);
+  this._entities['><'] = 0; // Dummy entry, so the first actual key is non-zero
   // `_entitiesInverse` is the inverse mapping of `_entities`, this is to avoid expensive inverse lookups.
   this._entitiesInverse = Object.create(null);
   this._entityCount = 0;
@@ -90,7 +91,7 @@ N3Store.prototype = {
   // Finally, `graph` will be the graph of the created triples.
   _findInIndex: function (index0, key0, key1, key2, name0, name1, name2, graph) {
     var vars = (key0 ? 1 : 0) + (key1 ? 1 : 0) + (key2 ? 1 : 0);
-    var results = [], entityKeys = (vars > 1 && !graph) ? Object.keys(this._entities) : this._entitiesInverse,
+    var results = [], entityKeys = vars > 1 ? Object.keys(this._entities) : this._entitiesInverse,
         tmp, index1, index2;
 
     // If a key is specified, use only that part of index 0.