--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -90,7 +90,7 @@ N3Store.prototype = {
   // (for instance: _subject_, _predicate_, and _object_).
   // Finally, `graph` will be the graph of the created triples.
   _findInIndex: function (index0, key0, key1, key2, name0, name1, name2, graph) {
-    var vars = (key0 ? 1 : 0) + (key1 ? 1 : 0) + (key2 ? 1 : 0);
+    var vars = (key0 ? 0 : 1) + (key1 ? 0 : 1) + (key2 ? 0 : 1);
     var results = [], entityKeys = vars > 1 ? Object.keys(this._entities) : this._entitiesInverse,
         tmp, index1, index2;