I fixed the issues from #61, lookups are significantly faster in all cases now, at the cost of a 1.2% increase in insertion times.
```
dim = 256
OLD:
    N3Store performance test
    - Adding 16777216 triples in the default graph: 20985.810ms
    * Memory usage for triples: 814MB
    - Finding all 16777216 triples to the default graph 65536 times (0 variables): 303956.190ms
    - Finding all 16777216 triples to the default graph 131072 times (1 variable): 8910.851ms
    - Finding all 16777216 triples to the default graph 196608 times (2 variables): 4105.836ms

    - Adding 16777216 quads: 27680.169ms
    * Memory usage for quads: 648MB
    - Finding all 16777216 quads 1048576 times: 20307.015ms
NEW:
    N3Store performance test
    - Adding 16777216 triples in the default graph: 21238.570ms
    * Memory usage for triples: 734MB
    - Finding all 16777216 triples to the default graph 65536 times (0 variables): 101509.681ms
    - Finding all 16777216 triples to the default graph 131072 times (1 variable): 4243.945ms
    - Finding all 16777216 triples to the default graph 196608 times (2 variables): 3401.100ms

    - Adding 16777216 quads: 26400.740ms
    * Memory usage for quads: 796MB
    - Finding all 16777216 quads 1048576 times: 19886.337ms
```